requirements.txt
django 
djangorestframework 
django-cors-headers
librosa 
torch 
torchaudio 
numpy 
pydub 
soundfile
